[Third-Party Client Hint Delegation] Add iframe wpts
Previously, we didn't test whether client hints were delegated to
iframes in addition to subresource requests. Now we can see that the
http-equiv tag doesn't modify permissions for iframes but the meta-name
tag does. The http-equiv tag probably should be modifying permissions
for same-origin frames to share client hints, but there's no reason to
fix the legacy behavior when we can have people move to the new tag.

Bug: 1219359
Change-Id: I3f92518417f2f48146f1691a01e4c04f6aca5471