[FlexNG] Remove baseline TODO
I wasn't sure if we needed to do something special for baselines and
fragmentation. It doesn't look like grid does anything special here,
so the TODO can be removed.

When looking at what grid was doing, I noticed there was a test for
baseline alignment, so I've added one, as well, (although this test
is unrelated to the TODO being removed).

Bug: 660611
Change-Id: Ifa50e3e6aa315515db6c1461e3e6bebafbd6c2a0